// src/onec/types/ut.create-data.ts
export interface IUtData {
  bidId: string;
  data: string;
  payDocNumber: string;
  sailerInn: string;
  sailerKpp: string;
  sailerBoxId: string;
  sailerOrgIdGuid: string;
  sailerFnsParticipantId: string;
  bayerInn: string;
  bayerKpp: string;
  bayerOrgIdGuid: string;
  bayerBoxId: string;
  bayerFnsParticipantId: string;
}